Allow user to retrieve all values from a CSV Line. (#16)
* csv: csv-lines: Allow user to retrieve all values from a CSV Line.

This change exposes the Line method in the sealed class that implements
ICsvLine, so the users can extract all values from a particular row. The
extracted values are going to be returned in string format, so we leave
the user the option to cast/parse values when needed instead of being
complex and clever to identify the type of a particular element.

This is a common operation when someone/something deals with CSV Files.
Sometimes it's necessary to have a collection of values for a particular
row(s).